requirements.txt
python-constraint
pandas
pygad
numpy